requirements.txt
django
MySQL-python==1.2.5
requests==2.3.0
pygithub
jenkinsapi
